fixed discussion points for alternative DNS answer - fixed strings - removed trailing empty lines in Slots.kt - moved generation of DNS answer to separated function - additional documentation
