+++ app/src/full/kotlin/tunnel/BlockaTunnelFiltering.kt
@@ -117,17 +117,7 @@ internal class BlockaTunnelFiltering(
         } else {
             dnsMessage.header.setFlag(Flags.QR.toInt())
             dnsMessage.header.rcode = Rcode.NOERROR
-            if(get(DnsAnswerState::class.java).hostNotFoundAnswer){
-                dnsMessage.addRecord(denyResponse, Section.AUTHORITY)
-            }else{
-                dnsMessage.addRecord(
-                    ARecord(Name(dnsMessage.question.name.toString(false)),
-                    DClass.IN,
-                    60 * 3,
-                    InetAddress.getByAddress(byteArrayOf(127, 1, 1, 1))),
-                    Section.ANSWER
-                )
-            }
+            generateDnsAnswer(dnsMessage, denyResponse)
             toDeviceFakeDnsResponse(dnsMessage.toWire(), originEnvelope)
             emit(TunnelEvents.REQUEST, Request(host, blocked = true))
             true
