--- app/src/full/kotlin/tunnel/DnsAnswerState.kt
@@ -1,6 +1,0 @@
-package tunnel
-
-
-data class DnsAnswerState(
-        val hostNotFoundAnswer: Boolean
-)
