+++ app/src/ui-blokada/kotlin/core/bits/Slots.kt
@@ -1318,6 +1318,3 @@ class CleanupVB(
         }
     }
 }
-
-
-
