+++ app/src/ui-blokada/res/values/strings_panel.xml
@@ -265,7 +265,7 @@
     <string name="slot_reset_counter_action">Reset</string>
 
 
-    <string name="label_workarounds">device-specific workarounds</string>
-    <string name="slot_dns_answer_label"> alternative DNS answer</string>
+    <string name="label_workarounds">Device-specific workarounds</string>
+    <string name="slot_dns_answer_label">Alternative DNS answer</string>
     <string name="slot_dns_answer_description">Changes the answer Blokada sends for blocked requests from "host not found" to an invalid ip. This can help on some ROMs which query other DNS servers if they get a "host not found" answer. please note that this might slow down some apps which try to connect to the invalid address.</string>
     </resources>
